Nested routes have been deprecated, move all of the routing to within the Application component's render function.  Create a function getCookie to simplify grabbing the csrftoken.
