+++ eventkit_cloud/ui/static/ui/app/components/Application.tsx
@@ -2,7 +2,6 @@ import * as PropTypes from 'prop-types';
 import * as React from 'react';
 import { connect } from 'react-redux';
 import axios from 'axios';
-import { withCookies, Cookies } from 'react-cookie';
 import { withTheme, withStyles, createStyles } from '@material-ui/core/styles';
 import withWidth, { isWidthUp } from '@material-ui/core/withWidth';
 import { Breakpoint } from '@material-ui/core/styles/createBreakpoints';
@@ -17,12 +16,36 @@ import { DrawerTimeout } from '../actions/uiActions';
 import {login, userActive} from '../actions/userActions';
 import { getNotifications, getNotificationsUnreadCount } from '../actions/notificationsActions';
 import NotificationsDropdown from './Notification/NotificationsDropdown';
+import Loadable from 'react-loadable';
+import { connectedReduxRedirect } from 'redux-auth-wrapper/history4/redirect';
+import { Route, RouteComponentProps } from 'react-router';
+import { routerActions } from 'connected-react-router';
+import PageLoading from './common/PageLoading';
 import '../styles/bootstrap/css/bootstrap.css';
 import '../styles/openlayers/ol.css';
 import '../styles/flexboxgrid.css';
 import '../styles/react-joyride-compliled.css';
 // tslint:disable-next-line:no-var-requires
 require('../fonts/index.css');
+
+const Loading = (args) => {
+    if (args.pastDelay) {
+        return (
+            <PageLoading background="pattern" />
+        );
+    }
+
+    return null;
+};
+
+function allTrue(acceptedLicenses) {
+    return Object.keys(acceptedLicenses).every(license => acceptedLicenses[license]);
+}
+
+const loadableDefaults = {
+    loading: Loading,
+    delay: 1000,
+};
 
 const jss = (theme: any) => createStyles({
     appBar: {
@@ -93,6 +116,7 @@ interface Props {
             identification: string;
         };
     };
+    isLoading: boolean;
     autoLogoutAt: Date;
     autoLogoutWarningAt: Date;
     notificationsData: Eventkit.Store.NotificationsData;
@@ -111,7 +135,6 @@ interface Props {
         notificationsIndicator: string;
     };
     dispatch: (options?: object) => void;
-    cookies: Cookies;
 }
 
 interface State {
@@ -129,6 +152,87 @@ interface State {
     loggedIn: boolean;
     isMounted: false;
 }
+
+const Loader = Loadable({
+    ...loadableDefaults,
+    loader: () => import('./common/PageLoading'),
+});
+
+const UserIsAuthenticated = connectedReduxRedirect({
+    authenticatedSelector: (props: Props) => !!props.userData,
+    authenticatingSelector: (props: Props) => props.isLoading,
+    AuthenticatingComponent: Loader,
+    redirectAction: routerActions.replace,
+    wrapperDisplayName: 'UserIsAuthenticated',
+    redirectPath: '/login',
+});
+
+const UserIsNotAuthenticated = connectedReduxRedirect({
+    redirectAction: routerActions.replace,
+    wrapperDisplayName: 'UserIsNotAuthenticated',
+    authenticatedSelector: (props: Props) => !props.userData && props.isLoading === false,
+    redirectPath: (state, ownProps: RouteComponentProps<{}, {}>) => (
+        ownProps.location.search || '/dashboard'
+    ),
+    allowRedirectBack: false,
+});
+
+const UserHasAgreed = connectedReduxRedirect({
+    redirectAction: routerActions.replace,
+    redirectPath: '/account',
+    wrapperDisplayName: 'UserHasAgreed',
+    authenticatedSelector: (props: Props) => allTrue(props.userData.accepted_licenses),
+});
+
+const LoginPage = Loadable({
+    ...loadableDefaults,
+    loader: () => import('./auth/LoginPage'),
+});
+
+const Logout = Loadable({
+    ...loadableDefaults,
+    loader: () => import('../containers/logoutContainer'),
+});
+
+const About = Loadable({
+    ...loadableDefaults,
+    loader: () => import('./About/About'),
+});
+
+const Account = Loadable({
+    ...loadableDefaults,
+    loader: () => import('./AccountPage/Account'),
+});
+
+const DashboardPage = Loadable({
+    ...loadableDefaults,
+    loader: () => import('./DashboardPage/DashboardPage'),
+});
+
+const DataPackPage = Loadable({
+    ...loadableDefaults,
+    loader: () => import('./DataPackPage/DataPackPage'),
+});
+
+const CreateExport = Loadable({
+    ...loadableDefaults,
+    loader: () => import('./CreateDataPack/CreateExport'),
+});
+
+const StatusDownload = Loadable({
+    ...loadableDefaults,
+    loader: () => import('./StatusDownloadPage/StatusDownload'),
+});
+
+const UserGroupsPage = Loadable({
+    ...loadableDefaults,
+    loader: () => import('./UserGroupsPage/UserGroupsPage'),
+});
+
+const NotificationsPage = Loadable({
+    ...loadableDefaults,
+    loader: () => import('./NotificationsPage/NotificationsPage'),
+});
 
 export class Application extends React.Component<Props, State> {
     private userActiveInputTypes: string[];
@@ -206,23 +310,17 @@ export class Application extends React.Component<Props, State> {
         return this.state.childContext;
     }
 
-     // getDerivedStateFromProps(props, state) {
-    //     // ...
-    // }
     checkAuth() {
+        console.log("userData", this.props.userData);
         if (!this.props.userData) {
-            const { cookies } = this.props;
-            const csrftoken = cookies.get('csrftoken');
-            this.props.dispatch(login(null, csrftoken));
+            this.props.dispatch(login(null));
         };
     }
 
     componentDidMount() {
         this.getConfig();
         this.checkAuth();
-        // this.setState({"isMounted": true}, () => {
-        //
-        // });
+        console.log("Props", this.props)
     }
 
     componentDidUpdate(prevProps, prevState) {
@@ -623,6 +721,19 @@ export class Application extends React.Component<Props, State> {
                 />
                 <div className="qa-Application-content" style={styles.content}>
                     <div>{childrenWithContext}</div>
+                    <Route path="/login" component={UserIsNotAuthenticated(LoginPage)} />
+                    <Route path="/logout" component={Logout} />
+                    <Route path="/dashboard" component={UserIsAuthenticated(UserHasAgreed(DashboardPage))} />
+                    <Route path="/exports" component={UserIsAuthenticated(UserHasAgreed(DataPackPage))} />
+                    <Route path="/create" component={UserIsAuthenticated(UserHasAgreed(CreateExport))} />
+                    <Route
+                        path="/status/:jobuid"
+                        component={UserIsAuthenticated(UserHasAgreed(StatusDownload))}
+                    />
+                    <Route path="/about" component={UserIsAuthenticated(About)} />
+                    <Route path="/account" component={UserIsAuthenticated(Account)} />
+                    <Route path="/groups" component={UserIsAuthenticated(UserGroupsPage)} />
+                    <Route path="/notifications" component={UserIsAuthenticated(NotificationsPage)} />
                 </div>
                 <BaseDialog
                     show={this.state.showAutoLogoutWarningDialog}
@@ -640,6 +751,7 @@ export class Application extends React.Component<Props, State> {
                     <strong>You have been automatically logged out due to inactivity.</strong>
                 </BaseDialog>
             </div>
+
         );
     }
 }
@@ -648,6 +760,7 @@ function mapStateToProps(state) {
     return {
         drawer: state.drawer,
         userData: state.user.data,
+        isLoading: state.user.status.isLoading,
         autoLogoutAt: state.user.meta.autoLogoutAt,
         autoLogoutWarningAt: state.user.meta.autoLogoutWarningAt,
         notificationsStatus: state.notifications.status,
@@ -680,10 +793,10 @@ function mapDispatchToProps(dispatch) {
 
 export default withWidth()(
     withTheme()<any>(
-        withStyles<any, any>(jss)(withCookies(
+        withStyles<any, any>(jss)(
             connect(mapStateToProps, mapDispatchToProps)(
                 Application
             )
         )
     )
-));
+);
