+++ eventkit_cloud/ui/static/ui/app/store/middlewares.ts
@@ -1,5 +1,4 @@
 import axios from 'axios';
-import { Cookies } from 'react-cookie';
 
 export const crashReporter = () => next => (action) => {
     try {
